// types.ts
export type UserRole = 'STUDENT' | 'TEACHER' | 'ADMIN' | 'SYSTEM';